Extract DEM grid city object projection
